GET request works.  added server.php and code for GET request
Remember, to run this, we need to copy it to the MAMP htdocs folder (or
symlink from there to my folder), start MAMP, and then access the
index.htm file from http://localhost:8888/ajax_test/index.htm.

I added the server.php file which receives the user-entered text,
converts it to uppercase, and then responds with it.

Then added the code in index.htm which sends the user submitted text,
and then takes the responded text (uppercase text) and displays it in
the other read-only text box.