Implemented GET/POST using jQuery
Added 2 more text fields with corresponding functions to do the
equivalent GET and POST but this time implemented using jQuery instead
of straight Javascript.  It’s nice because it’s a few less lines of
code, but even nicer because of the simple change to change between GET
and POST methods in jQuery.